requirements.txt
python-telegram-bot==21.4
openai==1.38.0
python-dotenv==1.0.1
googletrans==4.0.0-rc1
translate==3.6.1
httpx==0.27.0